Extract out chat logic from Store to specific module
